validator_structure: moved exceptions
