+++ microdata_tools/packaging/_encrypt.py
@@ -10,7 +10,7 @@ from cryptography.hazmat.primitives import hashes
 from cryptography.hazmat.primitives import serialization
 from cryptography.hazmat.primitives.asymmetric import padding
 
-from microdata_tools.exceptions import ValidationException
+from microdata_tools.packaging.exceptions import ValidationException
 from microdata_tools.packaging._utils import check_exists
 
 logger = logging.getLogger()
